Hangman
A Hangman game which generate a random animal name and then show blank spaces for each character of name. It allows user  to enter a character which if present in the name of the animal showed at its respective place. If user find all characters and complete the name of the animal, User Wins. If user enter a wrong character, one life is deducted and Hangman begins to complete. There are total of 6 chances. For each chance, there is a stage of Hangman. With each deducted chance, Hangman's initial stages will appear and eventually Hangman completes with 6th life deducted. In this case user lost.